pyproject.toml
[project]
name = "langgraph-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-openai>=1.2.1",
    "langgraph>=1.1.10",
    "python-dotenv>=1.2.2",
]
